move type_constructor to utils.yaml.py
